cleanup

+++ titan/mediathek/localparser_secret/streamcloud.sh
@@ -455,16 +455,6 @@ kino()
 #			TITLE=$(echo $ROUND | sed -nr "s/.*> <span title='([^']+)style=.*/\1/p")
 			TITLE=$(echo $ROUND | sed "s!<span title='!\nfound=!g" | sed "s!' style=!\n!g" | grep ^found= | grep -v href | cut -d"=" -f2 | head -n1)
 
-#			TITLE=`echo $ROUND | sed 's!title=!\nfound=>!g' | grep ^found= | cut -d'"' -f2 | tail -n1 | sed 's/ kostenlos//'`
-#			NEWPAGE=`echo $ROUND | sed 's!<a href=!\nfound=!g' | grep ^found= | cut -d'"' -f2 | tail -n1`
-
-#echo $ROUND
-#echo $ID
-#echo $PIC
-#echo $TITLE
-#echo $NEWPAGE
-
-#exit
 			if [ -z "$PIC" ]; then
 				PIC="http://atemio.dyndns.tv/mediathek/menu/default.jpg"
 			fi
@@ -489,16 +479,13 @@ kino()
 
 hosterlist()
 {
-rm $TMP/$FILENAME.list
-#	if [ ! -e "$TMP/$FILENAME.list" ]; then
+	if [ ! -e "$TMP/$FILENAME.list" ]; then
 		$curlbin "$PAGE" -o $TMP/cache.$FILENAME.1
 		cat $TMP/cache.$FILENAME.1 | grep iframe | sed -nr 's/.*src="([^"]+)".*/\1/p' >$TMP/cache.$FILENAME.2
 
 		while read -u 3 ROUND; do
-#echo ROUND $ROUND
 			NEWPAGE="$ROUND"
 			TITLE=`echo $ROUND | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
-#echo TITLE $TITLE
 			$curlbin "$NEWPAGE" -o $TMP/cache.$FILENAME.4
 
 #			<a class='play_container' href='http://www.vodlocker.to/embed?t=Die+Eisk%C3%B6nigin%3A+Olaf+taut+auf&y=2017&lang=de&referrer=link'>
@@ -508,51 +495,22 @@ rm $TMP/$FILENAME.list
 				TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
 				NEWPAGE=$URL
 			fi
-#echo URL2 $URL
-#echo TITLE2 $TITLE
 
-if [ ! -z "$URL" ];then
-			$curlbin "$URL" -o $TMP/cache.$FILENAME.5
-
-#			$("#hostname").html("<a href='http://openload.co/embed/vq1HFMJ5vAo/detroit.SD-spectre.mkv' target='_blank' style='color:gold; text-decoration:underline;'>openload.co</a>");
-			cat $TMP/cache.$FILENAME.5 | grep "#hostname" | sed -nr "s/.*a href='([^']+)'.*/\1/p" | grep -v "+link+" >$TMP/cache.$FILENAME.5.url1
-			URL=`cat $TMP/cache.$FILENAME.5.url1`
 			if [ ! -z "$URL" ];then
-				TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
-			fi 		
+				$curlbin "$URL" -o $TMP/cache.$FILENAME.5
 
-#echo URL3 $URL
-#echo TITLE3 $TITLE
+#				$("#hostname").html("<a href='http://openload.co/embed/vq1HFMJ5vAo/detroit.SD-spectre.mkv' target='_blank' style='color:gold; text-decoration:underline;'>openload.co</a>");
+				cat $TMP/cache.$FILENAME.5 | grep "#hostname" | sed -nr "s/.*a href='([^']+)'.*/\1/p" | grep -v "+link+" >$TMP/cache.$FILENAME.5.url1
+				URL=`cat $TMP/cache.$FILENAME.5.url1`
+				if [ ! -z "$URL" ];then
+					TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
+				fi 		
 
-#if [ ! -z "$URL" ];then
-#			if [ -z "$URL" ];then
-#
-#				cat $TMP/cache.$FILENAME.5 | sed 's/<source src=/\nfound=/g' | grep ^found= | cut -d"'" -f2 >$TMP/cache.$FILENAME.5.url2
-#				URL=`cat $TMP/cache.$FILENAME.5.url2`
-#				TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
-#			fi
-#echo URL4 $URL
-#echo TITLE4 $TITLE
-			if [ ! -z "$URL" ];then
-				TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
-				NEWPAGE=$URL
+				if [ ! -z "$URL" ];then
+					TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
+					NEWPAGE=$URL
+				fi
 			fi
-#fi
-
-		#echo 111111111
-#			if [ -z "$URL" ];then
-		#echo 22222222222
-#				cat $TMP/cache.$FILENAME.4 | grep "#hostname" | sed -nr "s/.*a href='([^']+)'.*/\1/p" >$TMP/cache.$FILENAME.5.url3
-#				URL=`cat $TMP/cache.$FILENAME.5.url3`
-
-#				TITLE=`echo $URL | sed -nr 's/.*:\/\/([^\/]+)\/.*/\1/p'` 		
-
-#				if [ ! -z "$URL" ];then
-#					/tmp/localhoster/hoster.sh get "$URL" > $TMP/cache.hoster.$hoster.url3
-#					URL=`cat $TMP/cache.hoster.$hoster.url3`
-#				fi
-#			fi
-fi
 
 			if [ ! -z "$TITLE" ] && [ "$TITLE" != " " ] && [ ! -z "$NEWPAGE" ];then
 				PIC=`echo $TITLE | tr [A-Z] [a-z] | sed 's/www.//' | cut -d"." -f1 | sed 's/streamclou/streamcloud/'`
@@ -560,26 +518,15 @@ fi
 
 				echo "$LINE" >> $TMP/$FILENAME.list
 			fi
-#echo $PIC
-#echo $TITLE
-#echo $NEWPAGE
 		done 3<$TMP/cache.$FILENAME.2
-#		rm $TMP/cache.$FILENAME.* > /dev/null 2>&1
-#	fi
+		rm $TMP/cache.$FILENAME.* > /dev/null 2>&1
+	fi
 	echo "$TMP/$FILENAME.list"
 }
 
 
 hoster()
 {
-	rm $TMP/cache.$FILENAME.* > /dev/null 2>&1
-#	$curlbin $URL/$PAGE -o $TMP/cache.$PARSER.$INPUT.1 -A 'Mozilla/5.0 (Windows NT 6.3; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Maxthon/4.4.7.3000 Chrome/30.0.1599.101 Safari/537.36'
-#	/tmp/localhoster/hoster.sh get $URL/$PAGE > $TMP/cache.$FILENAME.1
-
-
-#	/tmp/localhoster/hoster.sh get "$PAGE" > $TMP/cache.$FILENAME.1
-
-#	STREAMURL=`cat $TMP/cache.$FILENAME.1`
 	STREAMURL="$PAGE"
 	echo $STREAMURL
 }
@@ -590,13 +537,13 @@ case $INPUT in
 	mainmenu) $INPUT;;
 	hosterlist) $INPUT;;
 	hoster) $INPUT;;
-	search) $INPUT;;
-	searchtv) $INPUT;;
-	season) $INPUT;;
-	episode) $INPUT;;
+	search_disable) $INPUT;;
+	searchtv_disable) $INPUT;;
+	season_disable) $INPUT;;
+	episode_disable) $INPUT;;
 	kino) $INPUT;;
-	sorted) $INPUT;;
-	genre) $INPUT;;
-	sortedtv) $INPUT;;
-	genretv) $INPUT;;
+	sorted_disable) $INPUT;;
+	genre_disable) $INPUT;;
+	sortedtv_disable) $INPUT;;
+	genretv_disable) $INPUT;;
 esac
